signal-pipeline/requirements.txt
numpy>=1.26.0,<2.0
scipy>=1.12.0
pandas>=2.2.0
h5py>=3.10.0
pywt>=1.5.0
pyarrow>=15.0.0
scikit-learn>=1.4.0
matplotlib>=3.8.0
pytest>=8.0.0
pytest-cov>=4.1.0
pyyaml>=6.0.1
